v1.1.0(3 Days Left)
- Added SFX on items
- Added Sound effects for weapon abilities.
- SFX healings
- SFX powerups
- Bug fixed on inventory (stacking of items), added delay to prevent this
- double right-click to delete object
- Updated art for special effects.
- Adjust game difficulty and stages
- Added transition from pages
- Added In-Game Inventory, to allow players view what they collected and to delete items they don't want.
- Added music transition whenever entering new stages.

+++ Cattler/Assets/1.MainProj/Level/Level.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Level", menuName = "Levels/Level")]
 public class Level : ScriptableObject
 {
-    public int level; //the game stage
+    public float level; //the game stage
 
     ///Normal Enemies///
     //Type of Enemies
